+++ visualization/app/codeCharta/ui/dialog/dialog.service.ts
@@ -1,14 +1,9 @@
-import { dialogUrlParameterComponent } from "./dialog.urlParameter.component"
 import { dialogDownloadComponent } from "./dialog.download.component"
 import { dialogGlobalSettingsComponent } from "./dialog.globalSettings.component"
 
 export class DialogService {
 	/* @ngInject */
 	constructor(private $mdDialog) {}
-
-	public showQueryParamDialog() {
-		this.showCustomDialog(dialogUrlParameterComponent)
-	}
 
 	public showDownloadDialog() {
 		this.showCustomDialog(dialogDownloadComponent)