+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 
 - Removes Options panel from sidebar #528
 - Removes Weblinks panel from sidebar #528
+- Removed URL-parameter info from sidebar #525
 
 ### Fixed